Create an RTC example with automatic date and time setting
This example sets the RTC (Real Time Clock) on the Portenta C33
automatically by retrieving the date and time from the computer
you upload the sketch from, at the point when you start the upload.
Next, it gets the current time from the RTC and prints it to the
Serial Monitor. It then sets an RTC alarm to fire every time the
seconds value of the time is zero. The alarm, which now goes off
once a minute, triggers a callback that prints the current time to
the Serial Monitor.


Former-commit-id: c51128e776ddbcb0081b8bbe9104b4fe18031ca3