Fix freeze on open() after close() on SD card
It's necessary to save the assigned IRQs, or they will be lost and
replaced with FSP_INVALID_VECTOR if the SDCardBlockDevice is destroyed
and created again. If we don't, r_sdhi_card_identify() inside the Renesas FSP will
return FSP_ERR_RESPONSE because of a timeout when it tries to put the card in
idle mode, and R_SDHI_MediaInit() will fail.


Former-commit-id: 238d310eaaac3934c7b5ab99d521984446063409