Update action dependencies of GitHub Actions workflows
The repository's GitHub Actions workflows use several GitHub Actions actions. The steps that use the actions pin the
dependency to a major version series. This means the specific dependency version is unmanaged within that major version
series, but the reference must be updated after there is a new major version release of the dependency.

Several of the dependencies were pinned to outdated major version refs, causing the workflows to miss the benefits of
recent development work on those dependencies. They are hereby updated to the latest major version refs. The breaking
changes that caused the major version bumps in the dependencies are not relevant to our usage of them so no additional
adjustments are needed in order to perform these bumps.