WIP - SerialX(__UART__) tx speedup and fixes
This is a WIP.

The released code does not use the write(buf, cnt) code, but instead unrolls it to use the one byte at a time, as the method signature did not match that in print (missing const)

Fixing this uncovered lots of issues where data would be lost as, you would typically call to output again while the last one was still putting and could lose one or two bytes.

This code bypasses the R_SCI_UART_Write(&uart_ctrl, c, len); functions and instead I do my own code wheich utilizes the txBuffer that was already part of the class.  The code both fills the buffer and if the hardware buffer has room, it takes characters from the txBuffer to fill it.

I also implemented my own version of the TDRE interrupt, to again simply take data from our txBuffer.

But to do this without changing the underlying layers, I
implemented a version of attachInterruptVector, that plugs in the location of my ISR handler into the interrupt vector.

Probably could/should overwrite some of the others as well.

the SerialX.flush() code was completely busted.  But you hardly noticed it before as you were only outputting one character at a time and then waiting for an interrupt before returning.

I have some debug code still in here, hopefully disabled by a #define.