Bounds checking on g_pin_cfg array
As I mentioned in the forum thread:
https://forum.arduino.cc/t/apis-like-digitalwrite-who-use-g-pinc-cfg-should-do-bounds-checking/1156322

I believe that many of the simple functions should have some form of parameter testing.  For example: pinMode(100, OUTPUT);
Should fail instead of trying to use random garbage off the end of the array to pass down to the next level.

As @per1234 mentioned on the forum thread.  This has bounced around for years:
https://github.com/arduino/ArduinoCore-avr/pull/302

So decided to at least try to do it for a few of the APIs that have this issue.  Most of the other references to this array appear to either check or are driven by pin information in definded in the variant.