Add method to allow the GPT Timer period not to be buffered
During code review of changes I made to the Servo library:
https://github.com/arduino-libraries/Servo/pull/116

@iabdalkader requested that I move all of the platform specific timer changes into the core instead of being in the hardware specific sub-directory of the Servo library.

This change adds a method to tell the GPT timer, that when I make a change to the pseriod (set_period(p) ) that I want the change to happen now and not be buffered.

And updated the set_period method to check for this and directly set the not buffered register with the new period.