Display useful message when Linux post-install script has insufficient permissions
The Arduino Boards Manager automatically executes the `post_install.sh` script during installation of the platform on a
Linux machine.

This platform's post-install script is intended to create a udev rules file that gives write permissions for the USB
devices of the platform's boards. These permissions are required in order to upload to the boards.

The creation of the udev rules file requires superuser privileges, which are typically not available in the context of
the post-install script's execution. The script contains code to check whether the necessary privileges are available.
If not, it prints a message and skips the udev rules file creation.

Previously the message printed when the udev rules file creation was not possible was "Please run as root". This message
was completely meaningless to the user when printed during the Boards Manager installation. Worse, it might cause them
to think they must run the Arduino development software as root user, which is a bad idea and also wouldn't result in
the udev rules file being created since the installation of the platform to the user's account was already completed.

The message is hereby updated to provide a meaningful explanation of the potential problem as well as the specific
command the user can run from the terminal to execute the script as superuser.