FreeRTOS: allow override startup hooks default implementation
The default behaviour is do nothing: scheduler has to be started manually by the user.
Use AUTOSTART_FREERTOS if you declare tasks in your setup() and use them in loop().
Use EARLY_AUTOSTART_FREERTOS if you need the scheduler to be already running in setup().

Libraries that need to use threads/semaphores during setup() (for example, to implement the usual

myObj.begin(); <- this starts a thread
while (!myObj) {  <- this becomes true when the thread is unlocked
  delay();
}

will need to declare EARLY_START_FREERTOS_HOOK in one of their cpp files