Replace broken URLs in library metadata
The `url` property in the library metadata is used to direct the user to resources related to the library. The URL
defined by this property is used as the target for the "More info" link in Arduino Cloud Editor's Library Manager.

Previously, some of the platform bundled libraries had an incorrect `url` property value that only led to a 404 page.